+++ src/bot/constants/callback_data.py
@@ -6,4 +6,3 @@ GET_CATEGORIES = "categories_callback"
 ADD_CATEGORIES = "add_categories"
 CONFIRM_CATEGORIES = "confirm_categories"
 MENU = "menu_callback"
-TEST_EMAIL = "test_mailing"
