+++ src/bot/handlers/menu.py
@@ -1,19 +1,14 @@
-import json
-
 import structlog
-from telegram import InlineKeyboardButton, InlineKeyboardMarkup, ReplyKeyboardRemove, Update
+from telegram import Update
 from telegram.constants import ParseMode
 from telegram.ext import Application, CallbackQueryHandler, CommandHandler, ContextTypes, MessageHandler
 from telegram.ext.filters import StatusUpdate
 
 from src.bot.constants import callback_data, commands, enum, patterns
-from src.bot.handlers.form import FeedbackModel
 from src.bot.keyboards import get_back_menu, get_menu_keyboard, get_no_mailing_keyboard
 from src.bot.services.user import UserService
 from src.bot.utils import delete_previous_message
 from src.core.logging.utils import logger_decor
-from src.core.services.email import EmailProvider
-from src.settings import settings
 
 log = structlog.get_logger()
 
@@ -73,37 +68,6 @@ async def reason_handler(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
 
 @logger_decor
-async def web_app_data_handler(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    user_data = json.loads(update.effective_message.web_app_data.data)
-    buttons = [
-        [InlineKeyboardButton(text="Открыть меню", callback_data=callback_data.MENU)],
-        [InlineKeyboardButton(text="Посмотреть открытые задания", callback_data=callback_data.VIEW_TASKS)],
-    ]
-    keyboard = InlineKeyboardMarkup(buttons)
-    await update.message.reply_text(
-        text=f"Спасибо, я передал информацию команде ProCharity! Ответ придет на почту {user_data['email']}",
-        reply_markup=ReplyKeyboardRemove(),
-    )
-    await update.message.reply_text(
-        text="Вы можете вернуться в меню или посмотреть открытые задания. Нажмите на нужную кнопку.",
-        reply_markup=keyboard,
-    )
-    email_provider = EmailProvider()
-    form = FeedbackModel(
-        telegram_id=update.effective_user.id,
-        surname=user_data['surname'],
-        name=user_data['name'],
-        email=user_data['email'],
-        feedback=user_data['feedback']
-    )
-    await email_provider.send_question_feedback(
-        telegram_id=update.effective_user.id,
-        message=form.to_message(),
-        email=settings.EMAIL_ADMIN,
-    )
-
-
-@logger_decor
 @delete_previous_message
 async def about_project(update: Update, context: ContextTypes.DEFAULT_TYPE):
     await context.bot.send_message(
@@ -120,29 +84,9 @@ async def about_project(update: Update, context: ContextTypes.DEFAULT_TYPE):
     )
 
 
-@logger_decor
-async def test_mailing(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    """Проверка, что админам приходит электронное письмо на почту."""
-    await context.bot.send_message(
-        chat_id=update.effective_chat.id,
-        text="Админы, проверьте свою почту",
-        reply_markup=await get_back_menu(),
-        parse_mode=ParseMode.HTML,
-        disable_web_page_preview=True,
-    )
-    email_provider = EmailProvider()
-    await email_provider.send_question_feedback(
-        telegram_id=update.effective_user.id,
-        message="Тест - тут будет вопрос/отзыв от пользователя",
-        email=settings.EMAIL_ADMIN,
-    )
-
-
 def registration_handlers(app: Application):
     app.add_handler(CommandHandler(commands.MENU, menu_callback))
     app.add_handler(CallbackQueryHandler(menu_callback, pattern=callback_data.MENU))
     app.add_handler(CallbackQueryHandler(about_project, pattern=callback_data.ABOUT_PROJECT))
-    app.add_handler(MessageHandler(StatusUpdate.WEB_APP_DATA, web_app_data_handler))
     app.add_handler(CallbackQueryHandler(set_mailing, pattern=callback_data.JOB_SUBSCRIPTION))
     app.add_handler(CallbackQueryHandler(reason_handler, pattern=patterns.NO_MAILING_REASON))
-    app.add_handler(CallbackQueryHandler(test_mailing, pattern=callback_data.TEST_EMAIL))
